+++ RELEASES.md
@@ -7,6 +7,7 @@
 
 ### Fixed
  * wattsup-libftdi, wattsup-libusb: Fix empty reads by always setting baud rate and serial line properties
+ * wattsup-dev: Use pselect to avoid timer expired errors
 
 
 ## [v0.2.3] - 2017-11-19